Moves StrandList into Components folder
